requirements.txt
gunicorn==19.9.0
google-cloud-storage==1.13.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-WTF==0.14.2
glob2==0.6
h5py==2.8.0
html5lib==0.9999999
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
matplotlib==3.0.2
numpy==1.16.4
numpydoc==0.8.0
opencv-python-headless==4.1.0.25
pandas==0.23.4
pandocfilters==1.4.2
Pillow==5.3.0
py==1.7.0
regex==2018.1.10
requests==2.21.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
virtualenv==16.6.2
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
WTForms==2.2.1